Completed first draft of ColumnDefs. Need to have option for updates
