--- src/lib/GridFunctions/Crud/MetaData/README
@@ -1,1 +1,0 @@
-Make changes based on server side rules.
