+++ src/lib/GridParser/ColumnDefs/CellClassRules/index.js
@@ -14,10 +14,15 @@ const type_check = require('../../TypeCheck')
 function CellClassRulesInit( grid_column, is_editable, validator_function ) {
 
     cellClassRules = {}
-    cellClassRules['editable_pass_style'] = EditablePassStyle(is_editable, validator_function)
-    cellClassRules['editable_error_style'] = EditableErrorStyle(is_editable, validator_function)
-    cellClassRules['non_editable_pass_style'] = NonEditablePassStyle(is_editable, validator_function)
-    cellClassRules['non_editable_error_style'] = NonEditableErrorStyle(is_editable, validator_function)
+    if (type_check.IsNull(validator_function)) {
+        if (is_editabe) { cellClassRules['editable_pass_style'] = params => true } 
+        else { cellClassRules['non_editable_pass_style'] = params => true }
+    } else {
+        cellClassRules['editable_pass_style'] = EditablePassStyle(is_editable, validator_function)
+        cellClassRules['editable_error_style'] = EditableErrorStyle(is_editable, validator_function)
+        cellClassRules['non_editable_pass_style'] = NonEditablePassStyle(is_editable, validator_function)
+        cellClassRules['non_editable_error_style'] = NonEditableErrorStyle(is_editable, validator_function)
+    }
     grid_column['cellClassRules'] = cellClassRules
 
 }
